Add Data validation

+++ api/Controllers/CommentController.cs
@@ -24,6 +24,10 @@ namespace api.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
             var comments = await _commentRepository.GetAllAsync();
 
             var commentDto = comments.Select(s => s.ToCommentDto());
@@ -32,9 +36,13 @@ namespace api.Controllers
         }
 
         [HttpGet]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
             var comment = await _commentRepository.GetByIdAsync(id);
 
             if (comment == null)
@@ -46,9 +54,13 @@ namespace api.Controllers
         }
 
 
-        [HttpPost("{stockId}")]
+        [HttpPost("{stockId:int}")]
         public async Task<IActionResult> Create([FromRoute] int stockId, [FromBody] CreateCommentDto commentDto)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
             if (!await _stockRepository.StockExists(stockId))
             {
                 return BadRequest("Stock does not exist");
@@ -61,9 +73,13 @@ namespace api.Controllers
         }
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateDto)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
             var comment = await _commentRepository.UpdateAsync(id, updateDto.ToCommentFromUpdate());
 
             if (comment == null)
@@ -75,9 +91,13 @@ namespace api.Controllers
         }
 
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
             var commentModel = await _commentRepository.DeleteAsync(id);
 
             if (commentModel == null)
