+++ api/Controllers/StockController.cs
@@ -27,6 +27,10 @@ namespace api.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
             var stocks = await _stockRepository.GetAllAsync();
 
             var stockDto = stocks.Select(s => s.ToStockDto());
@@ -34,9 +38,13 @@ namespace api.Controllers
             return Ok(stocks);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
             var stock = await _stockRepository.GetByIdAsync(id);
 
             if (stock == null)
@@ -50,6 +58,10 @@ namespace api.Controllers
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
             var stockModel = stockDto.ToStockFromCreateDto();
 
             await _stockRepository.CreateAsync(stockModel);
@@ -58,9 +70,13 @@ namespace api.Controllers
         }
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto updateStockRequestDto)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
             var stockModel = await _stockRepository.UpdateAsync(id, updateStockRequestDto);
 
             if (stockModel is null)
@@ -72,9 +88,13 @@ namespace api.Controllers
         }
 
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
             var stockModel = await _stockRepository.DeleteAsync(id);
 
             if (stockModel is null)
